pyproject.toml
[project]
name = "imagefilter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.7",
    "pillow>=12.0.0",
    "requests>=2.32.5",
]
